+++ mowl/walking/node2vec/model.py
@@ -14,6 +14,8 @@ logger = logging.getLogger("node2vec")
 class Node2Vec(WalkingModel):
 
     '''
+    Implementation of DeepWalk based on [grover2016]_.
+
     :param p: Return hyperparameter. Default is 1.
     :type p: float
     :param q: In-out hyperparameter. Default is 1.