+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [0.3.0]
+### Added
+- Implemented `CategoricalProjector` based on [https://arxiv.org/abs/2305.07163](https://arxiv.org/abs/2305.07163). ([#59][i59])
+
+### Removed
+- Removed redundant class `based_models.EmbeddingModel`
+
+
 ## [0.2.1]
 ### Fixed
 Fixed issue related to importing graph-based models due to missing `__init__.py` files. ([#60][i60])
@@ -78,7 +86,9 @@ Fixed issue related to importing graph-based models due to missing `__init__.py`
 - Walking methods accept optional `outfile` parameter and corpus extraction methods do not append by default.
 - Documentation updated and fixed some typos.
 
-[unreleased]: https://github.com/bio-ontology-research-group/mowl/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/bio-ontology-research-group/mowl/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/bio-ontology-research-group/mowl/releases/tag/v0.3.0
+[0.2.1]: https://github.com/bio-ontology-research-group/mowl/releases/tag/v0.2.1
 [0.2.0]: https://github.com/bio-ontology-research-group/mowl/releases/tag/v0.2.0
 [0.1.1]: https://github.com/bio-ontology-research-group/mowl/releases/tag/v0.1.1
 [0.1.0]: https://github.com/bio-ontology-research-group/mowl/releases/tag/v0.1.0
@@ -89,4 +99,6 @@ Fixed issue related to importing graph-based models due to missing `__init__.py`
 [i36]: https://github.com/bio-ontology-research-group/mowl/issues/36
 [i42]: https://github.com/bio-ontology-research-group/mowl/issues/42
 [i43]: https://github.com/bio-ontology-research-group/mowl/issues/43
-[i43]: https://github.com/bio-ontology-research-group/mowl/issues/60
+[i59]: https://github.com/bio-ontology-research-group/mowl/issues/59
+[i60]: https://github.com/bio-ontology-research-group/mowl/issues/60
+